# Proper handling $crate Take 2 [DO NOT MERGE]