@matklad 2021-01-06
Couple of quick observations before review:

* during expansion, there are a lot of contiguous token ranges. We should take advantage of that, and store ranges, and not individual tokens in TokenMap
* I am unsure it we should store `TextRange`s in frames. It seems that, after origenal conversion of user-written syntax to mbe, we only need mappings between token ids