@edwin0cheng 2021-01-07
> per_query_memory_usage needs to account for new query.

I added `HygieneFrameQuery` in `per_query_memory_usage` .

> during expansion, there are a lot of contiguous token ranges. We should take advantage of that, and store ranges, and not individual tokens in TokenMap

IIUC, we only store token range in `TokenMap` right now.

> I am unsure it we should store TextRanges in frames. It seems that, after origenal conversion of user-written syntax to mbe, we only need mappings between token ids

We store the `TextRange` for the def site and call site syntax node, it is needed because at the current frame, we can't determine the upper level token id without knoning the text range of upper level syntax text starting position.

However, I changed it to just store the starting position (which is `TextSize`), it should save some memory.