@edwin0cheng 2021-01-07
Memory used : 

```
Database loaded 643.90ms, 284mi, 81mb
Crates in this dir: 36
Total modules found: 577
Total declarations: 11187
Total functions: 8742
Item Collection: 14.57s, 84873mi, 458mb
Total expressions: 241372
Expressions of unknown type: 2643 (1%)
Expressions of partially unknown type: 2076 (0%)
Type mismatches: 871
Inference: 28.91s, 135567mi, 633mb
Total: 43.48s, 220440mi, 1091mb
   243mb ItemTreeQuery
    76mb BodyQuery
    71mb HygieneFrameQuery
    57mb InferQueryQuery
    48mb TraitImplsInDepsQuery
    46mb TraitSolveQuery (purge)
    41mb CrateDefMapQueryQuery
    38mb ImplDatumQuery
    38mb FileTextQuery (purge)
    33mb MacroArgTextQuery
    28mb BodyWithSourceMapQuery
    26mb ImplDataQuery
...
```