@matklad 2021-01-07
yeah, that's quite high there. But we can't really not do this, so I think it's better to merge and optimize